# Gex.NetCore/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gex",
    "description": "Colleción de API's correspondientes al sistema de exámenes Gex.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@gexsystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/materia": {
      "get": {
        "tags": [
          "Materia"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseICollectionGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseICollectionGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseICollectionGexResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Materia"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Materia"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MateriaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Materia"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/materia/{id}": {
      "get": {
        "tags": [
          "Materia"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MateriaResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Object": {
        "type": "object",
        "nullable": true
      },
      "GexErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Object"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Object"
        },
        "nullable": true
      },
...
      "MateriaTipo": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "MateriaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "$ref": "#/components/schemas/MateriaTipo"
          }
        },
        "nullable": true
      },
...
      "MateriaResponseICollectionGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MateriaResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
      "PreguntaTipo": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "PreguntaRequest": {
        "required": [
          "descripcion",
          "id",
          "tema",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tema": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "tipo": {
            "$ref": "#/components/schemas/PreguntaTipo"
          },
          "materia_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "examen_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "nullable": true
      },
      "MateriaRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "$ref": "#/components/schemas/MateriaTipo"
          },
          "preguntas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreguntaRequest"
            },
            "nullable": true
          }
        },
        "nullable": true
      },
      "MateriaResponseGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MateriaResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
...
    }
  }
}